move service to server module
