+++ server/src/main/java/GroceryFamily/GroceryMom/service/mapper/PriceMapper.java
@@ -1,4 +1,4 @@
-package GroceryFamily.GroceryElders.service.mapper;
+package GroceryFamily.GroceryMom.service.mapper;
 
 import java.time.Instant;
 import java.util.ArrayList;
